requirements.txt
Pandas
numpypip 
joblib
scikit-learn 
Pillow
Django
django-allauth
djangorestframework
djangorestframework-queryfields
django-upgrade

